feat: add simple sample and full scan sample
